requirements.txt
Flask==3.0.3
mysql-connector-python==8.1.0
requests
selenium
beautifulsoup4==4.10.0